move chart component to components folder, remove types file
Load directly json
