--- pages/embryo/types.ts
@@ -1,6 +1,0 @@
-
-export type EmbryoLandingPageData = {
-  primaryViability: Array<{ label: string; value: number }>;
-  primaryViabilityChartData: Array<{ label: string; value: number }>;
-  windowsOfLethality: Array<{ label: string; value: number }>;
-}
